refactor(useFetchProducts): 폴더 구조 수정 및 테스트 케이스 파일 병합
